requirements.txt
torchvision~=0.26.0+cu126
pip~=25.2
attrs~=25.3.0
distro~=1.9.0
protobuf~=5.29.3
filelock~=3.19.1
streamlit~=1.51.0
pandas~=2.3.1
pillow~=10.4.0
torch~=2.11.0+cu126
optuna~=4.8.0
scikit-learn~=1.7.1